Created wrapper class for FAB.
This wrapper class handles both the FAB and the progress bar. The initial implementation sets the tint of the FAB icon to make the icon white, and has functions for displaying/hiding the progress bar. In the future this will also include animation functions for changing the icon when loading is complete.