Fixed requesting the UPass.
The checkbox input on the UPassBC page has a hidden element beside it, holding the value "false". Mechanize was adding this false value to the value array of the checkbox element when constructing the parameters to pass with the POST request, causing the Translink server to think we did not check any box, even when we did. Not sure why this hidden element exists, I suspect it might be an intricacy to make some JS magic work that Mechanize can't understand, hence the issue.

To fix it, remove this hidden element before submitting, so we don't add the value to the params in the request.

Hallelujah!