Properly throw authentication failed exception.
Previously the exception was thrown when we verified the page we were on was not the UPassBC page, however if you enter the wrong credentials it shouldn't even get that far. Instead, the page won't have the redirect forms, and so we will get an index exception when we try and get the form from the translinkRedirect. This case is now being handled, and thrown as an authentication exception.