Remove shadow on status bar
The App Bar was creating a shadow on the Status Bar in Main Activity, due to the wrong theme being selected in the Android Manifest. Fixing this and tweaking the theme to use the AppCompat.Light.NoActionBar theme (adhering to the Material Design Support Library tutorial here: https://android-developers.googleblog.com/2014/10/appcompat-v21-material-design-for-pre.html) fixed the issue.