Use WakefulBroadcastReceiver to prevent service from being killed
(This commit message is for the previous commit) The background renewal service was being killed in some instances before finishing. Since Android has gotten more strict on background services, a service now has to be started from a WakefulBroadcastReceiver (and be started using the completeWakefulIntent method) to ensure it isn't killed before its finished.